test_db_connection.py
from flask import Flask
from flask_sqlalchemy import SQLAlchemy
from sqlalchemy import text
import os
from dotenv import load_dotenv

load_dotenv()

def test_db_connection():
    app = Flask(__name__)
    
    # AWS RDS configuration
    RDS_HOST = os.getenv("RDS_HOST", "gaadimech-crm-db.cnewyw0y0leb.ap-south-1.rds.amazonaws.com")
    RDS_DB = os.getenv("RDS_DB", "crmportal")
    RDS_USER = os.getenv("RDS_USER", "postgres")
    RDS_PASSWORD = os.getenv("RDS_PASSWORD", "GaadiMech2024!")
    RDS_PORT = os.getenv("RDS_PORT", "5432")
    
    DATABASE_URL = f"postgresql+psycopg2://{RDS_USER}:{RDS_PASSWORD}@{RDS_HOST}:{RDS_PORT}/{RDS_DB}"
    
    app.config['SQLALCHEMY_DATABASE_URI'] = DATABASE_URL
    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
    
    db = SQLAlchemy(app)
    
    try:
        # Test the connection
        with app.app_context():
            result = db.session.execute(text('SELECT 1')).scalar()
            print("✅ Database connection successful!")
            
            # Test if tables exist
            tables = db.session.execute(text("""
                SELECT table_name 
                FROM information_schema.tables 
                WHERE table_schema = 'public'
            """)).fetchall()
            
            print("\nExisting tables:")
            for table in tables:
                print(f"- {table[0]}")
                
    except Exception as e:
        print(f"❌ Database connection failed: {str(e)}")
        return False
    
    return True

if __name__ == "__main__":
    test_db_connection() 