requirements.txt
# Flask and core dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Limiter==3.5.0

# Database drivers for PostgreSQL/RDS
psycopg2-binary==2.9.9
SQLAlchemy==2.0.41

# Security and environment
python-dotenv==1.0.0
Werkzeug==2.3.7

# Timezone handling
pytz==2023.3.post1

# Production server
gunicorn==21.2.0

# Additional utilities
requests==2.31.0

# Performance optimizations
redis==5.0.1
Flask-Caching==2.1.0

# Background tasks (if needed)
APScheduler==3.11.0

# Development tools (optional)
memory-profiler==0.61.0 